Remove Kabardian (Latin; kbd-latn)
It was added in 2010, when the Kabardian language
was first requested.[1]

The requester just asked for Kabardian (kbd), without
specifying the script. Later in the discussion, the requester
said that the writing in this language is usually in Cyrillic,
and kbd should be the primary code and use Cyrillic, and kbd-latn
should be separate.

Even though this language is occasionally written in
the Latin alphabet, I suggest removing it from
the translatewiki.net configuration for the following
reasons:
* No one ever really requested it.
* The currently defined autonym is almost certainly incorrect.
  It was probably copied from the English Wikipedia at the time,[2]
  but it was unreferenced, and since then it was replaced with
  something else, which is also unreferenced. It's unclear
  what should the correct Latin-script autonym be. I really
  prefer to be sure about the correctness of the language names
  we show on translatewiki.net.
* No translations have ever been made into this language
  on translatewiki.net,[3].
* I couldn't find any content in this language in the Latin
  script in any Wikimedia project—Wikipedia, Wikidata,
  Commons, etc. (The language code appears on a couple of
  technical pages in Commons as a possible parameter value,
  but there's no actual content, and the English Wikipedia
  had one name tagged as "kbd-latn" in a template, and I
  converted it to a transliteration parameter.)

Each one of these issues by itself is perhaps not a reason
to remove the language, but taken all together, it's
a case to remove.

I don't mind adding it back if a speaker of this language
who knows the Latin orthography asks for it.

[1] https://translatewiki.net/wiki/Thread:Project:Translator/_kbd_%D0%BA%D0%B0%D0%B1%D0%B0%D1%80%D0%B4%D0%B8%D0%BD%D1%81%D0%BA%D0%B8%D0%B9_%E2%80%93_kbd_(2)
[2] https://en.wikipedia.org/w/index.php?title=Kabardian_language&oldid=335474901
[3] https://translatewiki.net/wiki/Special:MessageGroupStats?group=core&messages=&x=D&uselang=en#sortable:2=asc

Change-Id: I89d153b8108edefc89eb54fef493e616a9b3b0a5