Update Cham languages configuration (cja, cjm)
I've been in contact with a speaker of this language
recently, so updating accordingly to that.

The autonyms for these two languages were written as requested,
but they were never referenced. The new ones are from this source:
Abdul Majid Hj. Yunos (2010). Kamus Cam–Melayu Jawi Rumi

In addition, only Latin script was configured,
even though the other scripts are used
for it, too, and there is no clearly dominant one.
So this patch adds the Arabic and Cham scripts.

Finally, it updates the default assistant languages.

Corresponding patch for language-data (and later, ULS):
https://github.com/wikimedia/language-data/pull/384

Change-Id: I13bf60bbc4449ca522273a8119fce2eb21cb1708