Fix rml language code
Follow-up to
If392d2aef1ca93d7f4744eef4a64f32f4a55727c

By mistake, I wrote "rml", but I meant to write "rml-cyrl"
in that patch.

The request was only for Cyrillic, which is used in Russia,
Belarus, and Ukraine. The Latin version, used in Poland,
appears to be quite different, so it should be added separately,
and only if there is a specific request for it.

Bug: T381838
Change-Id: I4d7ff106d2eafa07ba1f179a58eef66835d31293