Separate SPI icon from MediaWiki icon
Change the way the SPI icon is added, so it is not added to the same
div as the Powered by MediaWiki icon.

Visible change: On normal wiki pages, this leads to the Netcup
icon to appear *before* the other two icons (SPI, MediaWiki), whereas
before this change, the text (no icon before)  appeared *after* them.

Bug: T382753
Depends-On: Ia1bd743300836387663121677dfa727a5771aa14
Change-Id: I549f1cd894946cef36189ec5043df289145c0058