Increase edit rate limits for newbies
When editing via Special:Translate, each translation counts as
two edits for reasons unknown to me (see T386346), so the default
limit for newbies (= users who are not autoconfirmed) is
effectively halved from 8 to 4.

This patch increases the limit 4-fold from 8 (default) to 32, thus
increasing the effective limit from 8 to 16 assuming most of the
edits come via Special:Translate.

Bug: T386335
Change-Id: If45dc175f83eb84dd843acd7664fce5b377d654f