[OpenStreetMap] Alphabetize optional messages
This makes no functional changes, but only sorts
the keys alphabetically in order to make future
maintenance easier.

This also removes one comment that looks like
it's about one of the messages, but is actually
irrelevant because it was added very long ago about
another message, which is now unused, and was probably
left in the file by mistake.

Change-Id: I7864de2486243b28613eefc0e387abacec61f0d0