Remove validation exclusion for PLURAL/GENDER in MW messages
When {{GENDER:$1}} and {{PLURAL:$1}} were first added for many
messages, translators were complaining that they weren't
necessary for their languages, and as a result, those messages
were added to the validation exclusion list(s). However, over the
years, a standard has developed that even in languages that don't
need them, we still translate it as {{GENDER:$1|something}}, and
so we don't need to add validation exclusion for such messages.

Bug: T407135
Depends-On: Ifd184567464d64f48f7d0691b4c47025e2811c51
Depends-On: I16f90269ccc6f4435dd29db38fa94474a303885c
Change-Id: I376e1e558f0463e20eb645cea0b29f27aa0f1d54